refactor: split TUI model and render helpers
